+++ src/py_oidc_auth/__init__.py
@@ -63,7 +63,7 @@ if TYPE_CHECKING:
     from .quart_auth import QuartOIDCAuth
     from .tornado_auth import TornadoOIDCAuth
 
-__version__ = "2603.0.2"
+__version__ = "2604.0.0"
 
 _LAZY_IMPORTS = {
     "FastApiOIDCAuth": ".fastapi_auth",